requirements.txt
alembic==1.13.0
APIFlask==2.0.2
apispec==6.3.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-HTTPAuth==4.8.0
Flask-JWT-Extended==4.6.0
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
greenlet==3.0.2
idna==3.6
importlib-metadata==7.0.0
importlib-resources==6.1.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
marshmallow==3.20.1
packaging==23.2
pluggy==1.3.0
PyJWT==2.8.0
PyMySQL==1.1.0
pytest==7.4.3
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
responses==0.24.1
SQLAlchemy==2.0.23
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.1.0
webargs==8.3.0
Werkzeug==3.0.1
zipp==3.17.0
